pyproject.toml
[project]
name = "realtime-ga-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.7.0",
    "wandb>=0.17.1",
    "sentry-sdk>=2.0.0,<3.0.0",
    "pydantic>=2.0.0",
    "packaging>=21.0",                      # For version parsing in integrations
    "tenacity>=8.3.0,!=8.4.0",              # Excluding 8.4.0 because it had a bug on import of AsyncRetrying
    "click",                                # Used for terminal/stdio printing
    "gql[httpx]>=3.0.0",                    # Used exclusively in wandb_api.py; httpx is used for HTTP requests throughout the codebase
    "jsonschema>=4.23.0",                   # Used by scorers for field validation
    "diskcache==5.6.3",                     # Used for data caching
    "tzdata; platform_system == 'Windows'", # Required for timezone support on Windows
    "cachetools>=6.2.4",                    # Used for TTLCache in call_batch_processor
    # Used for content type detection
    # Remove once license content filtering is upstreamed
    "polyfile-weave>=0.5.9",
    "websockets>=16.0",
    "pyaudio>=0.2.14",
    "numpy>=2.4.2",
    "ddtrace>=2.7.0",
    "boto3>=1.34.0",
    "azure-storage-blob>=12.24.0,<12.26.0",
    "google-cloud-storage>=2.7.0",
    "litellm>=1.36.1",
    "opentelemetry-proto>=1.12.0",
    "opentelemetry-semantic-conventions-ai>=0.4.3",
    "openinference-semantic-conventions>=0.1.17",
    "websocket-client>=1.9.0",
]
